+++ test/core/di/reflective_provider.spec.ts
@@ -5,8 +5,8 @@ import { global } from '../../../src/facade/lang';
 import { Component, Directive } from '../../../src/core/directives/decorators';
 import { Injectable, Inject } from '../../../src/core/di/decorators';
 import { Pipe } from '../../../src/core/pipes/decorators';
-import { _getNgModuleMetadataByType, resolveReflectiveProvider } from '../../../src/core/di/reflective_provider';
-import { createNgModule } from '../../utils';
+import { _getAngular1ModuleMetadataByType, resolveReflectiveProvider } from '../../../src/core/di/reflective_provider';
+import { createAngular1Module } from '../../utils';
 import { getInjectableName } from '../../../src/core/di/provider';
 import { createProvider } from '../../../src/core/di/provider_util';
 import { OpaqueToken } from '../../../src/core/di/opaque_token';
@@ -21,7 +21,7 @@ describe( `di/reflective_provider`, () => {
   const sandbox = sinon.sandbox.create();
 
   beforeEach( () => {
-    global.angular = createNgModule() as any;
+    global.angular = createAngular1Module() as any;
   } );
   afterEach( () => {
     sandbox.restore();
@@ -41,7 +41,7 @@ describe( `di/reflective_provider`, () => {
 
     } );
 
-    it( `should resolve value by useValue and provide metadata for ngModule`, () => {
+    it( `should resolve value by useValue and provide metadata for angular1Module`, () => {
 
       const ValueToken = new OpaqueToken('helloToken');
       const FalseToken = new OpaqueToken('helloFalse');
@@ -61,7 +61,7 @@ describe( `di/reflective_provider`, () => {
       expect( actual ).to.deep.equal( expected );
 
     } );
-    it( `should resolve service by useClass and provide metadata for ngModule`, () => {
+    it( `should resolve service by useClass and provide metadata for angular1Module`, () => {
 
       const classToken = new OpaqueToken('classToken');
 
@@ -87,7 +87,7 @@ describe( `di/reflective_provider`, () => {
       expect( actual ).to.deep.equal( expected );
 
     } );
-    it( `should resolve factory by useFactory and provide metadata for ngModule`, () => {
+    it( `should resolve factory by useFactory and provide metadata for angular1Module`, () => {
 
       const classToken = new OpaqueToken('classToken');
 
@@ -125,7 +125,7 @@ describe( `di/reflective_provider`, () => {
       expect( actual[ 1 ].value.$inject ).to.deep.equal( [ '$log','$http' ] );
 
     } );
-    it( `should resolve factory by useExisting and provide metadata for ngModule`, () => {
+    it( `should resolve factory by useExisting and provide metadata for angular1Module`, () => {
 
       @Injectable()
       class SomeInjectable{}
@@ -139,16 +139,16 @@ describe( `di/reflective_provider`, () => {
   } );
   describe( `#_normalizeProviders`, () => {
 
-    let ngModule: ng.IModule;
+    let angular1Module: ng.IModule;
     beforeEach( () => {
-      ngModule = global.angular.module( 'myApp', [] );
+      angular1Module = global.angular.module( 'myApp', [] );
     } );
 
     it( `should add dependant module to existing one if provider is string`, () => {
       const provider = 'ui.bootstrap.datepicker';
-      const updatedNgModule = _normalizeProviders( ngModule, [provider] );
+      const updatedAngular1Module = _normalizeProviders( angular1Module, [provider] );
 
-      expect( updatedNgModule.requires ).to.deep.equal( [ 'ui.bootstrap.datepicker' ] );
+      expect( updatedAngular1Module.requires ).to.deep.equal( [ 'ui.bootstrap.datepicker' ] );
     } );
     it( `should register $provide via value/service/factory if provider is ProviderLiteral`, () => {
 
@@ -160,11 +160,11 @@ describe( `di/reflective_provider`, () => {
         { provide: 'myHelloSvc', useClass: HelloSvc },
         { provide: 'myHelloFactory', useFactory: ()=>new HelloSvc() }
       ];
-      const updatedNgModule = _normalizeProviders( ngModule, providers );
+      const updatedAngular1Module = _normalizeProviders( angular1Module, providers );
 
-      expect( _isTypeRegistered( 'myValToken', updatedNgModule, '$provide', 'value' ) ).to.equal( true );
-      expect( _isTypeRegistered( 'myHelloSvc', updatedNgModule, '$provide', 'service' ) ).to.equal( true );
-      expect( _isTypeRegistered( 'myHelloFactory', updatedNgModule, '$provide', 'factory' ) ).to.equal( true );
+      expect( _isTypeRegistered( 'myValToken', updatedAngular1Module, '$provide', 'value' ) ).to.equal( true );
+      expect( _isTypeRegistered( 'myHelloSvc', updatedAngular1Module, '$provide', 'service' ) ).to.equal( true );
+      expect( _isTypeRegistered( 'myHelloFactory', updatedAngular1Module, '$provide', 'factory' ) ).to.equal( true );
 
     } );
     it( `should register $provide via service if provider is Decorated Type`, () => {
@@ -172,9 +172,9 @@ describe( `di/reflective_provider`, () => {
       class HelloSvc{}
 
       const provider = HelloSvc;
-      const updatedNgModule = _normalizeProviders( ngModule, [provider] );
+      const updatedAngular1Module = _normalizeProviders( angular1Module, [provider] );
 
-      expect( _isTypeRegistered( getInjectableName(HelloSvc), updatedNgModule, '$provide', 'service' ) ).to.equal( true );
+      expect( _isTypeRegistered( getInjectableName(HelloSvc), updatedAngular1Module, '$provide', 'service' ) ).to.equal( true );
     } );
     it( `should register $compileProvider via directive if provider is Decorated Type`, () => {
 
@@ -185,18 +185,18 @@ describe( `di/reflective_provider`, () => {
       class MyComponent { }
 
       const providers = [ HelloAttrDirective, MyComponent ];
-      const updatedNgModule = _normalizeProviders( ngModule, [ providers ] );
+      const updatedAngular1Module = _normalizeProviders( angular1Module, [ providers ] );
 
       expect( _isTypeRegistered(
         getInjectableName( HelloAttrDirective ),
-        updatedNgModule,
+        updatedAngular1Module,
         '$compileProvider',
         'directive'
       ) ).to.equal( true );
 
       expect( _isTypeRegistered(
         getInjectableName( MyComponent ),
-        updatedNgModule,
+        updatedAngular1Module,
         '$compileProvider',
         'directive'
       ) ).to.equal( true );
@@ -209,11 +209,11 @@ describe( `di/reflective_provider`, () => {
       }
 
       const provider = UpsHelloPipe;
-      const updatedNgModule = _normalizeProviders( ngModule, [provider] );
+      const updatedAngular1Module = _normalizeProviders( angular1Module, [provider] );
 
       expect( _isTypeRegistered(
         getInjectableName( UpsHelloPipe ),
-        updatedNgModule,
+        updatedAngular1Module,
         '$filterProvider',
         'register'
       ) ).to.equal( true );
@@ -228,9 +228,9 @@ describe( `di/reflective_provider`, () => {
         return internalRef;
       }
 
-      const updatedNgModule = _normalizeProviders( ngModule, [ stateConfig, createProvider( {} ) ] ) as any;
+      const updatedAngular1Module = _normalizeProviders( angular1Module, [ stateConfig, createProvider( {} ) ] ) as any;
 
-      expect( updatedNgModule._configBlocks ).to.deep.equal(
+      expect( updatedAngular1Module._configBlocks ).to.deep.equal(
         [
           [ '$injector', 'invoke', [ stateConfig ] ],
           [ '$injector', 'invoke', [ internalRef ] ]
@@ -239,47 +239,47 @@ describe( `di/reflective_provider`, () => {
 
     } );
     it( `should throw if non supported provider type is used`, () => {
-      expect( ()=>_normalizeProviders( ngModule, [ 23213 ] as any ) ).to.throw();
-      expect( ()=>_normalizeProviders( ngModule, [ {} ] as any ) ).to.throw();
-      expect( ()=>_normalizeProviders( ngModule, [ true ] as any ) ).to.throw();
+      expect( ()=>_normalizeProviders( angular1Module, [ 23213 ] as any ) ).to.throw();
+      expect( ()=>_normalizeProviders( angular1Module, [ {} ] as any ) ).to.throw();
+      expect( ()=>_normalizeProviders( angular1Module, [ true ] as any ) ).to.throw();
     } );
 
   } );
   describe( `#_isTypeRegistered`, () => {
 
-    let ngModule: ng.IModule;
+    let angular1Module: ng.IModule;
 
     @Component({selector:'my-cmp',template:'fooo'})
     class MyCmp{}
 
     beforeEach( () => {
-      ngModule = global.angular.module( 'myApp', [] );
+      angular1Module = global.angular.module( 'myApp', [] );
     } );
 
-    it( `should check ngModule for duplicates`, () => {
+    it( `should check angular1Module for duplicates`, () => {
 
-      let actual = _isTypeRegistered( 'myValue', ngModule, '$provide', 'value' );
+      let actual = _isTypeRegistered( 'myValue', angular1Module, '$provide', 'value' );
 
       expect( actual ).to.equal( false );
 
-      ngModule.value( ...provide( 'myValue', { useValue: 'hello' } ) );
-      actual = _isTypeRegistered( 'myValue', ngModule, '$provide', 'value' );
+      angular1Module.value( ...provide( 'myValue', { useValue: 'hello' } ) );
+      actual = _isTypeRegistered( 'myValue', angular1Module, '$provide', 'value' );
 
       expect( actual ).to.equal( true );
 
-      actual = _isTypeRegistered( 'myCmp', ngModule, '$compileProvider', 'directive' );
+      actual = _isTypeRegistered( 'myCmp', angular1Module, '$compileProvider', 'directive' );
 
       expect( actual ).to.equal( false );
 
-      ngModule.directive( ...provide( MyCmp ) );
-      actual = _isTypeRegistered( 'myCmp', ngModule, '$compileProvider', 'directive' );
+      angular1Module.directive( ...provide( MyCmp ) );
+      actual = _isTypeRegistered( 'myCmp', angular1Module, '$compileProvider', 'directive' );
 
       expect( actual ).to.equal( true );
 
     } );
 
   } );
-  describe(`#_getNgModuleMetadataByType`, () => {
+  describe(`#_getAngular1MetadataByType`, () => {
 
     @Component({selector:'foo',template:'hello'})
     class FooComponent{}
@@ -305,11 +305,11 @@ describe( `di/reflective_provider`, () => {
         return configPhase;
       }
 
-      expect( () => _getNgModuleMetadataByType( Configure ) ).to.not.throw();
-      expect( () => _getNgModuleMetadataByType( configPhase ) ).to.not.throw();
-      expect( () => _getNgModuleMetadataByType( configFactory( {} ) ) ).to.not.throw();
+      expect( () => _getAngular1ModuleMetadataByType( Configure ) ).to.not.throw();
+      expect( () => _getAngular1ModuleMetadataByType( configPhase ) ).to.not.throw();
+      expect( () => _getAngular1ModuleMetadataByType( configFactory( {} ) ) ).to.not.throw();
 
-      expect( _getNgModuleMetadataByType( configPhase ) ).to.deep.equal( {
+      expect( _getAngular1ModuleMetadataByType( configPhase ) ).to.deep.equal( {
         providerName: '$injector',
         providerMethod: 'invoke',
         moduleMethod: 'config'
@@ -317,12 +317,12 @@ describe( `di/reflective_provider`, () => {
 
     });
 
-    it(`should return ngModule registration method by Type Metadata`, () => {
+    it(`should return angular1Module registration method by Type Metadata`, () => {
       const actual = [
-        _getNgModuleMetadataByType( FooComponent ),
-        _getNgModuleMetadataByType( FooDirective ),
-        _getNgModuleMetadataByType( MyService ),
-        _getNgModuleMetadataByType( UpsPipe ),
+        _getAngular1ModuleMetadataByType( FooComponent ),
+        _getAngular1ModuleMetadataByType( FooDirective ),
+        _getAngular1ModuleMetadataByType( MyService ),
+        _getAngular1ModuleMetadataByType( UpsPipe ),
       ];
       const expected = [
         { providerName: '$compileProvider', providerMethod: 'directive', moduleMethod: 'directive' },
@@ -351,39 +351,39 @@ describe( `di/reflective_provider`, () => {
 
     class JustClass{}
 
-    let ngModule: ng.IModule;
+    let angular1Module: ng.IModule;
 
     beforeEach( () => {
-      ngModule = global.angular.module( 'myApp', [] );
+      angular1Module = global.angular.module( 'myApp', [] );
     } );
 
     it( `should do nothing if Class doesn't have annotation`, () => {
 
-      _registerTypeProvider( ngModule, JustClass, { moduleMethod: 'service', name: '', value: noop } );
-      expect( (ngModule as any)._invokeQueue ).to.deep.equal( [] );
+      _registerTypeProvider( angular1Module, JustClass, { moduleMethod: 'service', name: '', value: noop } );
+      expect( (angular1Module as any)._invokeQueue ).to.deep.equal( [] );
 
     } );
     it( `should register component/pipe/service via its name`, () => {
 
-      sandbox.spy(ngModule,'service');
-      sandbox.spy(ngModule,'filter');
-      sandbox.spy(ngModule,'directive');
+      sandbox.spy(angular1Module,'service');
+      sandbox.spy(angular1Module,'filter');
+      sandbox.spy(angular1Module,'directive');
 
-      _registerTypeProvider( ngModule, MyService, { moduleMethod: 'service', name: 'myService#1', value: noop } );
-      _registerTypeProvider( ngModule, UpsPipe, { moduleMethod: 'filter', name: 'ups', value: noop } );
-      _registerTypeProvider( ngModule, FooComponent, { moduleMethod: 'directive', name: 'foo', value: noop } );
+      _registerTypeProvider( angular1Module, MyService, { moduleMethod: 'service', name: 'myService#1', value: noop } );
+      _registerTypeProvider( angular1Module, UpsPipe, { moduleMethod: 'filter', name: 'ups', value: noop } );
+      _registerTypeProvider( angular1Module, FooComponent, { moduleMethod: 'directive', name: 'foo', value: noop } );
 
-      expect( (ngModule.service as Sinon.SinonSpy).calledOnce ).to.equal( true );
-      expect( (ngModule.filter as Sinon.SinonSpy).calledOnce ).to.equal( true );
-      expect( (ngModule.directive as Sinon.SinonSpy).calledOnce ).to.equal( true );
+      expect( (angular1Module.service as Sinon.SinonSpy).calledOnce ).to.equal( true );
+      expect( (angular1Module.filter as Sinon.SinonSpy).calledOnce ).to.equal( true );
+      expect( (angular1Module.directive as Sinon.SinonSpy).calledOnce ).to.equal( true );
 
     } );
     it( `should register directive via 3 types of template usage (name),[name],name`, () => {
 
-      sandbox.spy( ngModule, 'directive' );
-      const spy = ngModule.directive as Sinon.SinonSpy;
+      sandbox.spy( angular1Module, 'directive' );
+      const spy = angular1Module.directive as Sinon.SinonSpy;
 
-      _registerTypeProvider( ngModule, FooDirective, { moduleMethod: 'directive', name: 'foo', value: noop } );
+      _registerTypeProvider( angular1Module, FooDirective, { moduleMethod: 'directive', name: 'foo', value: noop } );
 
       expect( spy.calledThrice ).to.equal( true );
       expect( spy.calledWith( 'foo', noop ) ).to.equal( true );